+++ sources/sound_engine/elements/enveloppevol.cpp
@@ -24,6 +24,7 @@
 
 #include "enveloppevol.h"
 #include "qmath.h"
+#include "voiceparam.h"
 
 void EnveloppeVol::initialize(quint32 sampleRate, bool isMod)
 {
@@ -54,8 +55,8 @@ bool EnveloppeVol::applyEnveloppe(float * data, quint32 size, bool release, int 
         v_timeDecay = voiceParam->getDouble(champ_decayModEnv) * _sampleRate;
         v_levelSustain = voiceParam->getDouble(champ_sustainModEnv);
         v_timeRelease = voiceParam->getDouble(champ_releaseModEnv) * _sampleRate;
-        v_noteToHold = (float)voiceParam->getInteger(champ_keynumToModEnvHold) / 1200;
-        v_noteToDecay = (float)voiceParam->getInteger(champ_keynumToModEnvDecay) / 1200;
+        v_noteToHold = static_cast<float>(voiceParam->getInteger(champ_keynumToModEnvHold) / 1200);
+        v_noteToDecay = static_cast<float>(voiceParam->getInteger(champ_keynumToModEnvDecay) / 1200);
     }
     else
     {
@@ -65,8 +66,8 @@ bool EnveloppeVol::applyEnveloppe(float * data, quint32 size, bool release, int 
         v_timeDecay = voiceParam->getDouble(champ_decayVolEnv) * _sampleRate;
         v_levelSustain = voiceParam->getDouble(champ_sustainVolEnv);
         v_timeRelease = voiceParam->getDouble(champ_releaseVolEnv) * _sampleRate;
-        v_noteToHold = (float)voiceParam->getInteger(champ_keynumToVolEnvHold) / 1200;
-        v_noteToDecay = (float)voiceParam->getInteger(champ_keynumToVolEnvDecay) / 1200;
+        v_noteToHold = static_cast<float>(voiceParam->getInteger(champ_keynumToVolEnvHold) / 1200);
+        v_noteToDecay = static_cast<float>(voiceParam->getInteger(champ_keynumToVolEnvDecay) / 1200);
     }
 
     if (_fastRelease)
@@ -85,8 +86,8 @@ bool EnveloppeVol::applyEnveloppe(float * data, quint32 size, bool release, int 
                 static_cast<float>(qPow(10, -0.05 * static_cast<double>(v_levelSustain))); // decrease in dB
 
     // Update hold / decay time and volume according to the key
-    quint32 timeHold = static_cast<quint32>(v_timeHold * fastPow2(v_noteToHold * (60 - note)));
-    quint32 timeDecay = static_cast<quint32>(v_timeDecay * fastPow2(v_noteToDecay * (60 - note)));
+    quint32 timeHold = static_cast<quint32>(v_timeHold * fastPow2(v_noteToHold * static_cast<float>(60 - note)));
+    quint32 timeDecay = static_cast<quint32>(v_timeDecay * fastPow2(v_noteToDecay * static_cast<float>(60 - note)));
 
     // Avancement
     bool end = false;
